refactor: ♻️ extractHeadings
